+++ git-updater.php
@@ -12,7 +12,7 @@
  * Plugin Name:       Git Updater
  * Plugin URI:        https://github.com/afragen/git-updater
  * Description:       A plugin to automatically update GitHub hosted plugins, themes, and language packs. Additional API plugins available for Bitbucket, GitLab, Gitea, and Gist.
- * Version:           9.9.12.1
+ * Version:           9.9.12.2
  * Author:            Andy Fragen
  * License:           MIT
  * Domain Path:       /languages
@@ -25,6 +25,8 @@
  */
 
 namespace Fragen\Git_Updater;
+
+use Fragen\Git_Updater\API\Zipfile_API;
 
 /*
  * Exit if called directly.
@@ -45,6 +47,7 @@ if ( ! class_exists( 'Fragen\Git_Updater\Bootstrap' ) ) {
 	( new Bootstrap( __FILE__ ) )->deactivate_die();
 }
 
+( new Zipfile_API() )->load_hooks();
 add_action(
 	'plugins_loaded',
 	function() {