+++ src/Git_Updater/Bootstrap.php
@@ -10,7 +10,7 @@
 
 namespace Fragen\Git_Updater;
 
-use Fragen\Git_Updater\PRO\Init as Init_PRO;
+use Fragen\Git_Updater\PRO\Bootstrap as Bootstrap_PRO;
 
 /*
  * Exit if called directly.
@@ -84,6 +84,10 @@ class Bootstrap {
 			return;
 		}
 
+		$this->freemius();
+		if ( gu_fs()->is__premium_only() ) {
+			( new Bootstrap_PRO() )->run();
+		}
 		( new Init() )->run();
 
 		register_activation_hook( $this->file, [ new Init(), 'rename_on_activation' ] );
@@ -134,6 +138,70 @@ class Bootstrap {
 		foreach ( $crons as $cron ) {
 			$timestamp = \wp_next_scheduled( $cron );
 			\wp_unschedule_event( $timestamp, $cron );
+		}
+	}
+
+	/**
+	 * Freemius integration.
+	 *
+	 * @return array|void
+	 */
+	public function freemius() {
+		if ( ! function_exists( 'gu_fs' ) ) {
+
+			/**
+			 * Create a helper function for easy SDK access.
+			 *
+			 * @return \stdClass
+			 */
+			function gu_fs() {
+				global $gu_fs;
+
+				if ( ! isset( $gu_fs ) ) {
+					// Activate multisite network integration.
+					if ( ! defined( 'WP_FS__PRODUCT_8195_MULTISITE' ) ) {
+						define( 'WP_FS__PRODUCT_8195_MULTISITE', true );
+					}
+
+					$gu_fs = fs_dynamic_init(
+						[
+							'id'               => '8195',
+							'slug'             => 'git-updater',
+							'premium_slug'     => 'git-updater',
+							'type'             => 'plugin',
+							'public_key'       => 'pk_2cf29ecaf78f5e10f5543c71f7f8b',
+							'is_premium'       => true,
+							'is_premium_only'  => true,
+							'has_addons'       => true,
+							'has_paid_plans'   => true,
+							'is_org_compliant' => false,
+							'trial'            => [
+								'days'               => 14,
+								'is_require_payment' => false,
+							],
+							'menu'             => [
+								'slug'    => 'git-updater',
+								'contact' => false,
+								'support' => false,
+								'network' => true,
+								'parent'  => [
+									'slug' => 'options-general.php',
+								],
+							],
+							// Set the SDK to work in a sandbox mode (for development & testing).
+							// IMPORTANT: MAKE SURE TO REMOVE SECRET KEY BEFORE DEPLOYMENT.
+							'secret_key'       => '',
+						]
+					);
+				}
+
+				return $gu_fs;
+			}
+
+			// Init Freemius.
+			gu_fs();
+			// Signal that SDK was initiated.
+			do_action( 'gu_fs_loaded' );
 		}
 	}
 }